Swapped form POST for fetch POST Added redirects after image upload Split UploadPopup.tsx from Popup.tsx Specified TODO Fixed typo in PupupBtn.tsx
