+++ src/pages/ProductDetails.tsx
@@ -8,7 +8,8 @@ export default function ProductDetails()  {
     const [image, setImage] = useState<IImage>(); //getter and setter / pobieram i ustawiem
     //wykonywany gdy ten komponent (ten plik na którym jesteśmy) zostanie wyswietlony na strony czyli zainicjalizowany lub zmieni się jakaś wartość w useState i use Effect to są Hooki()
     useEffect(() => {
-        if (image === undefined) {
+        console.log('ProductDetails - id is: '+id);
+        if (image === undefined || image._id != id) {
             fetch('http://localhost:3000/api/image/' + id)
                 .then(res => res.json())
 
@@ -17,26 +18,27 @@ export default function ProductDetails()  {
                     console.log(res)
                 });
         }
-    }, [image]);//jakie wartości będą nasłuchiwane
+    }, [image, id]);//jakie wartości będą nasłuchiwane
 
     if(image === undefined) {
         return (
             <p>Brak danych</p>
         );
     }
+    else{
+        return (
+            <div className='flex justify-center p-2'>
 
-    return (
-        <div className='flex justify-center p-2'>
+                <ImageFull
+                    key={image._id}
+                    imageSrc={image.imageBase64}
+                    title={image.title}
+                    description={image.description}
+                />
+                {}
 
-            <ImageFull
-                key={image._id}
-                imageSrc={image.imageBase64}
-                title={image.title}
-                description={image.description}
-            />
-            {/* <Image imageSrc={image.images[0]} imageAlt={image.category} /> */}
-
-        </div>
-    )
+            </div>
+        )
+    }
     
 }
